requirements.txt
langchain
langgraph
langchain-community
langchain-openai
tavily-python
langchain-tavily
python-dotenv
ruff
mypy